pyproject.toml
[project]
name = "astrovoice"
version = "1.0.0"
description = "AI-powered voice astrology consultation platform"
authors = [{name = "AstroVoice Team"}]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "openai>=1.3.7",
    "pydub>=0.25.1",
    "websockets>=12.0",
    "jinja2>=3.1.2",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "tabulate>=0.9.0",
    "aiohttp>=3.9.1",
    "Pillow>=10.4.0",
]

[project.optional-dependencies]
database = [
    "psycopg2-binary>=2.9.10",
    "boto3>=1.28.85",
]
deployment = [
    "mangum>=0.17.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["backend"]

[tool.setuptools.package-data]
backend = ["**/*.sql", "**/*.json"]

